rest of bias to ElementD

+++ tests/kernels/test_cutlass.py
@@ -85,7 +85,7 @@ def cutlass_int8_gemm_helper(m: int,
 
     if bias:
         # bias term should be > 1 so that the absolute tolerance can catch it
-        bias_t = torch.rand((n, ), device=device, dtype=torch.float32) + 1.0
+        bias_t = torch.rand((n, ), device=device, dtype=out_dtype) + 1.0
         out = ops.cutlass_scaled_mm(a, b, scale_a, scale_b, out_dtype, bias_t)
     else:
         out = ops.cutlass_scaled_mm(a, b, scale_a, scale_b, out_dtype)
